arrumando ex

+++ 17-08-4.c
@@ -7,21 +7,21 @@ int main(void)
 
     setlocale(LC_ALL, "Portuguese");
 
-    int x1, x2, y1, y2, F;
+    float x1, x2, y1, y2, F;
 
 
     printf("Coloque o X1: ");
-    scanf("%i",&x1);
+    scanf("%f",&x1);
     printf("Coloque o X2: ");
-    scanf("%i",&x2);
+    scanf("%f",&x2);
     printf("Coloque o y1: ");
-    scanf("%i",&y1);
+    scanf("%f",&y1);
     printf("Coloque o y2: ");
-    scanf("%i",&y2);
+    scanf("%f",&y2);
 
     F = sqrt(pow(x1-x2,2) + pow(y1-y2,2));
 
-    printf("A tersa parte deste número é: %i", F);
+    printf("A tersa parte deste número é: %f", F);
 
     return 0;
 }
